Updated new plugin tracker model
