+++ src/main/java/org/hydev/mcpm/client/database/ListAllTestCases.java
@@ -1,0 +1,10 @@
+package org.hydev.mcpm.client.database;
+
+public class TestLocalPluginTracker {
+    @Test
+    public void testReadMetadata() {
+        LocalPluginTracker localPluginTracker = new LocalPluginTracker();
+        localPluginTracker.readMeta();
+
+    }
+}
